refactor: move command bar composables to the composables folder
