+++ app/javascript/dashboard/composables/commands/helpers/actions.js
