+++ app/javascript/dashboard/composables/commands/helpers/events.js
