+++ app/javascript/dashboard/composables/commands/helpers/icons.js
