+++ app/javascript/dashboard/composables/commands/useBulkActionsHotKeys.js
@@ -7,15 +7,15 @@ import {
   CMD_BULK_ACTION_SNOOZE_CONVERSATION,
   CMD_BULK_ACTION_REOPEN_CONVERSATION,
   CMD_BULK_ACTION_RESOLVE_CONVERSATION,
-} from './commandBarBusEvents';
+} from './helpers/events';
 import {
   ICON_SNOOZE_CONVERSATION,
   ICON_REOPEN_CONVERSATION,
   ICON_RESOLVE_CONVERSATION,
-} from './CommandBarIcons';
+} from './helpers/icons';
 import { emitter } from 'shared/helpers/mitt';
 
-import { createSnoozeHandlers } from './commandBarActions';
+import { createSnoozeHandlers } from './helpers/actions';
 
 const SNOOZE_OPTIONS = wootConstants.SNOOZE_OPTIONS;
